requirements.txt
pandasgoogle-cloud-storagejoblibscikit-learn